Updates!
Added new UI element to the towerSpawnerBase ( the price of each tower is now correctly displayed). Also changed the logic regarding the checking of balance and withdrawal of coins when spawning a tower
Added a new type of tower (The catapult) - WIP, it will slow down enemies. Has everything except the finalized script for it. (It might need serious fine tuning later on).